+++ lib/ts/recipe/emailpassword/components/library/formBase.tsx
@@ -25,7 +25,7 @@ import STGeneralError from "supertokens-web-js/utils/error";
 import { MANDATORY_FORM_FIELDS_ID_ARRAY } from "../../constants";
 
 import type { APIFormField } from "../../../../types";
-import type { FormBaseProps } from "../../types";
+import type { FormBaseProps, FormFieldThemeProps } from "../../types";
 import type { FormEvent } from "react";
 
 import { Button, FormRow, Input, InputError, Label } from ".";
@@ -36,6 +36,89 @@ type FieldState = {
     error?: string;
     value: string;
 };
+
+const fetchDefaultValue = (field: FormFieldThemeProps): string => {
+    if (field.getDefaultValue !== undefined) {
+        const defaultValue = field.getDefaultValue();
+        if (typeof defaultValue !== "string") {
+            throw new Error(`getDefaultValue for ${field.id} must return a string`);
+        } else {
+            return defaultValue;
+        }
+    }
+    return "";
+};
+
+function InputComponentWrapper(props: {
+    field: FormFieldThemeProps;
+    type: string;
+    fstate: FieldState;
+    onInputFocus: (field: APIFormField) => void;
+    onInputBlur: (field: APIFormField) => void;
+    onInputChange: (field: APIFormField) => void;
+}) {
+    const { field, type, fstate, onInputFocus, onInputBlur, onInputChange } = props;
+
+    const useCallbackOnInputFocus = useCallback<(value: string) => void>(
+        (value) => {
+            onInputFocus({
+                id: field.id,
+                value,
+            });
+        },
+        [onInputFocus, field]
+    );
+
+    const useCallbackOnInputBlur = useCallback<(value: string) => void>(
+        (value) => {
+            onInputBlur({
+                id: field.id,
+                value,
+            });
+        },
+        [onInputBlur, field]
+    );
+
+    const useCallbackOnInputChange = useCallback(
+        (value) => {
+            onInputChange({
+                id: field.id,
+                value,
+            });
+        },
+        [onInputChange, field]
+    );
+
+    return field.inputComponent !== undefined ? (
+        <field.inputComponent
+            type={type}
+            name={field.id}
+            validated={fstate.validated === true}
+            placeholder={field.placeholder}
+            value={fstate.value}
+            autoComplete={field.autoComplete}
+            autofocus={field.autofocus}
+            onInputFocus={useCallbackOnInputFocus}
+            onInputBlur={useCallbackOnInputBlur}
+            onChange={useCallbackOnInputChange}
+            hasError={fstate.error !== undefined}
+        />
+    ) : (
+        <Input
+            type={type}
+            name={field.id}
+            validated={fstate.validated === true}
+            placeholder={field.placeholder}
+            value={fstate.value}
+            autoComplete={field.autoComplete}
+            onInputFocus={useCallbackOnInputFocus}
+            onInputBlur={useCallbackOnInputBlur}
+            onChange={useCallbackOnInputChange}
+            autofocus={field.autofocus}
+            hasError={fstate.error !== undefined}
+        />
+    );
+}
 
 export const FormBase: React.FC<FormBaseProps<any>> = (props) => {
     const { footer, buttonLabel, showLabels, validateOnBlur, formFields } = props;
@@ -50,7 +133,7 @@ export const FormBase: React.FC<FormBaseProps<any>> = (props) => {
     }, [unmounting]);
 
     const [fieldStates, setFieldStates] = useState<FieldState[]>(
-        props.formFields.map((f) => ({ id: f.id, value: "" }))
+        props.formFields.map((f) => ({ id: f.id, value: fetchDefaultValue(f) }))
     );
     const [isLoading, setIsLoading] = useState(false);
 
@@ -95,6 +178,9 @@ export const FormBase: React.FC<FormBaseProps<any>> = (props) => {
 
     const onInputChange = useCallback(
         (field: APIFormField) => {
+            if (typeof field.value !== "string") {
+                throw new Error(`${field.id} value must be a string`);
+            }
             updateFieldState(field.id, (os) => ({ ...os, value: field.value, error: undefined }));
             props.clearError();
         },
@@ -165,10 +251,18 @@ export const FormBase: React.FC<FormBaseProps<any>> = (props) => {
                     // If field error.
                     if (result.status === "FIELD_ERROR") {
                         const errorFields = result.formFields;
-
-                        setFieldStates((os) =>
-                            os.map((fs) => ({ ...fs, error: errorFields.find((ef: any) => ef.id === fs.id)?.error }))
-                        );
+                        const getErrorMessage = (fs: FieldState) => {
+                            const errorMessage = errorFields.find((ef: any) => ef.id === fs.id)?.error;
+                            if (errorMessage === "Field is not optional") {
+                                const fieldConfigData = props.formFields.find((f) => f.id === fs.id);
+                                // replace non-optional server error message from nonOptionalErrorMsg
+                                if (fieldConfigData?.nonOptionalErrorMsg !== undefined) {
+                                    return fieldConfigData?.nonOptionalErrorMsg;
+                                }
+                            }
+                            return errorMessage;
+                        };
+                        setFieldStates((os) => os.map((fs) => ({ ...fs, error: getErrorMessage(fs) })));
                     }
                 }
             } catch (e) {
@@ -191,12 +285,11 @@ export const FormBase: React.FC<FormBaseProps<any>> = (props) => {
                 if (field.id === "confirm-password") {
                     type = "password";
                 }
-                const fstate: FieldState = fieldStates.find((s) => s.id === field.id) || {
-                    id: field.id,
-                    validated: false,
-                    error: undefined,
-                    value: "",
-                };
+
+                const fstate: FieldState | undefined = fieldStates.find((s) => s.id === field.id);
+                if (fstate === undefined) {
+                    throw new Error("Should never come here");
+                }
 
                 return (
                     <FormRow key={field.id} hasError={fstate.error !== undefined}>
@@ -208,36 +301,14 @@ export const FormBase: React.FC<FormBaseProps<any>> = (props) => {
                                     <Label value={field.label} showIsRequired={field.showIsRequired} />
                                 ))}
 
-                            {field.inputComponent !== undefined ? (
-                                <field.inputComponent
-                                    type={type}
-                                    name={field.id}
-                                    validated={fstate.validated === true}
-                                    placeholder={field.placeholder}
-                                    value={fstate.value}
-                                    autoComplete={field.autoComplete}
-                                    autofocus={field.autofocus}
-                                    onInputFocus={onInputFocus}
-                                    onInputBlur={onInputBlur}
-                                    onChange={onInputChange}
-                                    hasError={fstate.error !== undefined}
-                                />
-                            ) : (
-                                <Input
-                                    type={type}
-                                    name={field.id}
-                                    validated={fstate.validated === true}
-                                    placeholder={field.placeholder}
-                                    value={fstate.value}
-                                    autoComplete={field.autoComplete}
-                                    onInputFocus={onInputFocus}
-                                    onInputBlur={onInputBlur}
-                                    onChange={onInputChange}
-                                    autofocus={field.autofocus}
-                                    hasError={fstate.error !== undefined}
-                                />
-                            )}
-
+                            <InputComponentWrapper
+                                type={type}
+                                field={field}
+                                fstate={fstate}
+                                onInputFocus={onInputFocus}
+                                onInputBlur={onInputBlur}
+                                onInputChange={onInputChange}
+                            />
                             {fstate.error && <InputError error={fstate.error} />}
                         </Fragment>
                     </FormRow>